bug: turn off transparency in Stata versions before 15
We use transparency to visually highlight point estimates relative to
confidence intervals. However, the color%alpha syntax returns an
unknown color in Stata 14 and older, since transparency was added in
Stata 15.

This should close #18.

This commit also contains a quality of life update to the CI plot option
logic, which was unnecessarily verbose and resulted in surprising
behavior when an invalid CI plot type was specified. Now rcap is always
the default.